refactor(service): GameService 인터페이스 패키지 분리한다.
