+++ src/main/kotlin/baseball/player/Player.kt
@@ -3,6 +3,5 @@ package baseball.player
 import camp.nextstep.edu.missionutils.Console
 
 class Player {
-
-    fun enterNumber(): String = Console.readLine()
+    fun enterNumber(): String = Console.readLine().trim()
 }
